pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Alexandra Goldobina <goldobina.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.10.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
isort = "^5.13.2"
ipython = "^8.28.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.4"
drf-spectacular = "^0.27.2"
stripe = "^11.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
